Move export win date utilities to local module
