+++ src/client/utils/date.js
@@ -7,21 +7,14 @@
  */
 
 const {
-  addDays,
-  differenceInDays,
   differenceInCalendarDays,
-  isSameDay,
   endOfToday,
-  startOfMonth: getStartOfMonth,
-  isWithinInterval,
   formatDistanceToNowStrict,
   isAfter,
   isValid,
   parse,
   parseISO,
-  subMonths,
   differenceInCalendarMonths,
-  isEqual: areDatesEqual,
 } = require('date-fns')
 
 const {
@@ -215,47 +208,6 @@ function convertUnparsedDateToFieldDateObject(date) {
   return { day: '', month: '', year: '' }
 }
 
-/**
- * Generates a random date within the range specified by startDate and endDate (inclusive).
- * @param {Date} startDate - The start date of the range.
- * @param {Date} endDate - The end date of the range.
- * @returns {Date} A random date within the specified range.
- * @throws {Error} If startDate is greater than endDate or if startDate and endDate are the same date.
- */
-function getRandomDateInRange(startDate, endDate) {
-  if (isSameDay(startDate, endDate)) {
-    throw new Error('Start date and end date cannot be the same.')
-  }
-  if (startDate > endDate) {
-    throw new Error('Start date cannot be greater than end date.')
-  }
-  const daysDifference = differenceInDays(endDate, startDate)
-  const randomNumberOfDays = Math.floor(Math.random() * (daysDifference + 1))
-  return addDays(startDate, randomNumberOfDays)
-}
-
-/**
- * Returns the start date (1st day) of the month twelve months ago from the current date.
- * @returns {Date} The start date (1st day) of the month twelve months ago from the current date.
- */
-function getStartDateOfTwelveMonthsAgo() {
-  return subMonths(getStartOfMonth(new Date()), 12)
-}
-
-/**
- * Checks if a given date falls within the last twelve months from the current date.
- * The last twelve months include the 1st of the month.
- * @param {Date} date - The date to be checked.
- * @returns {boolean} Returns true if the date falls within the last twelve months
- * from the current date, otherwise false.
- */
-function isWithinLastTwelveMonths(date) {
-  return isWithinInterval(date, {
-    start: getStartDateOfTwelveMonthsAgo(),
-    end: new Date(),
-  })
-}
-
 module.exports = {
   generateFinancialYearLabel,
   getDifferenceInDays,
@@ -267,11 +219,6 @@ module.exports = {
   formatStartAndEndDate,
   convertDateToFieldShortDateObject,
   convertDateToFieldDateObject,
-  getRandomDateInRange,
-  isWithinLastTwelveMonths,
-  getStartDateOfTwelveMonthsAgo,
-  getStartOfMonth,
-  areDatesEqual,
   convertUnparsedDateToFieldDateObject,
   convertUnparsedDateToFieldShortDateObject,
 }
